+++ lib/rabarber.rb
@@ -1,22 +1,23 @@
 # frozen_string_literal: true
 
 require_relative "rabarber/version"
-require_relative "rabarber/configuration"
 
 require "active_record"
 require "active_support"
 
-require_relative "rabarber/input/base"
-require_relative "rabarber/input/action"
-require_relative "rabarber/input/ar_model"
-require_relative "rabarber/input/authorization_context"
-require_relative "rabarber/input/context"
-require_relative "rabarber/input/dynamic_rule"
-require_relative "rabarber/input/role"
-require_relative "rabarber/input/roles"
-require_relative "rabarber/input/types/boolean"
-require_relative "rabarber/input/types/proc"
-require_relative "rabarber/input/types/symbol"
+require_relative "rabarber/inputs"
+
+require_relative "rabarber/configuration"
+
+module Rabarber
+  class Error < StandardError; end
+  class ConfigurationError < Rabarber::Error; end
+  class InvalidArgumentError < Rabarber::Error; end
+  class NotFoundError < Rabarber::Error; end
+
+  delegate :configure, to: Rabarber::Configuration
+  module_function :configure
+end
 
 require_relative "rabarber/core/cache"
 
@@ -32,15 +33,3 @@ require_relative "rabarber/core/permissions"
 require_relative "rabarber/core/integrity_checker"
 
 require_relative "rabarber/railtie"
-
-module Rabarber
-  class Error < StandardError; end
-  class ConfigurationError < Rabarber::Error; end
-  class InvalidArgumentError < Rabarber::Error; end
-  class NotFoundError < Rabarber::Error; end
-
-  def configure
-    yield(Rabarber::Configuration.instance)
-  end
-  module_function :configure
-end