+++ lib/rabarber/controllers/concerns/authorization.rb
@@ -21,11 +21,34 @@ module Rabarber
       def grant_access(action: nil, roles: nil, context: nil, if: nil, unless: nil)
         Rabarber::Core::Permissions.add(
           self,
-          Rabarber::Input::Action.new(action).process,
-          Rabarber::Input::Roles.new(roles).process,
-          Rabarber::Input::AuthorizationContext.new(context).process,
-          Rabarber::Input::DynamicRule.new(binding.local_variable_get(:if)).process,
-          Rabarber::Input::DynamicRule.new(binding.local_variable_get(:unless)).process
+          Rabarber::Inputs.process(
+            action,
+            as: :symbol,
+            optional: true,
+            message: "Expected a symbol or a string, got #{action.inspect}"
+          ),
+          Rabarber::Inputs.process(
+            roles,
+            as: :roles,
+            message: "Expected an array of symbols or strings containing only lowercase letters, numbers, and underscores, got #{roles.inspect}"
+          ),
+          Rabarber::Inputs.process(
+            context,
+            as: :authorization_context,
+            message: "Expected a Class, an instance of ActiveRecord model, a symbol, a string, or a proc, got #{context.inspect}"
+          ),
+          Rabarber::Inputs.process(
+            binding.local_variable_get(:if),
+            as: :dynamic_rule,
+            optional: true,
+            message: "Expected a symbol, a string, or a proc, got #{binding.local_variable_get(:if).inspect}"
+          ),
+          Rabarber::Inputs.process(
+            binding.local_variable_get(:unless),
+            as: :dynamic_rule,
+            optional: true,
+            message: "Expected a symbol, a string, or a proc, got #{binding.local_variable_get(:unless).inspect}"
+          )
         )
       end
     end