+++ lib/rabarber/models/role.rb
@@ -6,7 +6,7 @@ module Rabarber
 
     belongs_to :context, polymorphic: true, optional: true
 
-    has_and_belongs_to_many :roleables, class_name: Rabarber::Configuration.instance.user_model_name,
+    has_and_belongs_to_many :roleables, class_name: Rabarber::Configuration.user_model_name,
                                         association_foreign_key: "roleable_id",
                                         join_table: "rabarber_roles_roleables"
 
@@ -63,8 +63,7 @@ module Rabarber
       end
 
       def assignees(name, context: nil)
-        find_by(name: process_role_name(name), **process_context(context))&.roleables ||
-          Rabarber::Configuration.instance.user_model.none
+        find_by(name: process_role_name(name), **process_context(context))&.roleables || Rabarber::Configuration.user_model.none
       end
 
       private
@@ -76,11 +75,19 @@ module Rabarber
       end
 
       def process_role_name(name)
-        Rabarber::Input::Role.new(name).process
+        Rabarber::Inputs.process(
+          name,
+          as: :role,
+          message: "Expected a symbol or a string containing only lowercase letters, numbers, and underscores, got #{name.inspect}"
+        )
       end
 
       def process_context(context)
-        Rabarber::Input::Context.new(context).process
+        Rabarber::Inputs.process(
+          context,
+          as: :role_context,
+          message: "Expected an instance of ActiveRecord model, a Class, or nil, got #{context.inspect}"
+        )
       end
     end