+++ spec/rabarber/controllers/concerns/authorization_spec.rb
@@ -51,31 +51,31 @@ RSpec.describe Rabarber::Authorization do
     context "when action is invalid" do
       let(:args) { { action: 1 } }
 
-      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Action name must be a Symbol or a String"
+      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Expected a symbol or a string, got 1"
     end
 
     context "when roles are invalid" do
       let(:args) { { roles: "junior developer" } }
 
-      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Role names must be Symbols or Strings and may only contain lowercase letters, numbers, and underscores"
+      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Expected an array of symbols or strings containing only lowercase letters, numbers, and underscores, got \"junior developer\""
     end
 
     context "when context is invalid" do
       let(:args) { { context: 1 } }
 
-      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Context must be a Class, an instance of ActiveRecord model, a Symbol, a String, or a Proc"
+      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Expected a Class, an instance of ActiveRecord model, a symbol, a string, or a proc, got 1"
     end
 
     context "when dynamic rule is invalid" do
       let(:args) { { if: 1 } }
 
-      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Dynamic rule must be a Symbol, a String, or a Proc"
+      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Expected a symbol, a string, or a proc, got 1"
     end
 
     context "when negated dynamic rule is invalid" do
       let(:args) { { unless: 1 } }
 
-      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Dynamic rule must be a Symbol, a String, or a Proc"
+      it_behaves_like "raises an error", Rabarber::InvalidArgumentError, "Expected a symbol, a string, or a proc, got 1"
     end
 
     context "when everything is valid" do
@@ -84,15 +84,6 @@ RSpec.describe Rabarber::Authorization do
       it "adds the permission" do
         expect(Rabarber::Core::Permissions).to receive(:add)
           .with(DummyAuthController, :index, [:admin], { context_id: nil, context_type: nil }, :foo, :bar).and_call_original
-        subject
-      end
-
-      it "uses Inputs to process the arguments" do
-        { index: :Action, admin: :Roles, nil => :AuthorizationContext, foo: :DynamicRule, bar: :DynamicRule }.each do |arg, type|
-          input_processor = instance_double(Rabarber::Input.const_get(type))
-          allow(Rabarber::Input.const_get(type)).to receive(:new).with(arg).and_return(input_processor)
-          expect(input_processor).to receive(:process)
-        end
         subject
       end
     end