+++ spec/rabarber/controllers/shared_examples.rb
@@ -22,6 +22,6 @@ end
 shared_examples_for "it raises an error on nil current_user" do |hash|
   it "raises an error" do
     expect { send(hash.keys.first, hash.values.first, params: hash[:params]) }
-      .to raise_error(Rabarber::Error, "Expected `current_user` to return an instance of User, but got nil")
+      .to raise_error(Rabarber::Error, "Expected `current_user` to return an instance of User, got nil")
   end
 end