+++ spec/rabarber/models/concerns/has_roles_spec.rb
@@ -14,19 +14,8 @@ RSpec.describe Rabarber::HasRoles do
     it "raises an error when the given roles are invalid" do
       expect { subject }.to raise_error(
         Rabarber::InvalidArgumentError,
-        "Role names must be Symbols or Strings and may only contain lowercase letters, numbers, and underscores"
+        "Expected an array of symbols or strings containing only lowercase letters, numbers, and underscores, got [:Admin, \"junior developer\"]"
       )
-    end
-  end
-
-  shared_examples_for "role names are processed" do
-    let(:roles) { [:admin, :manager] }
-
-    it "uses Input::Roles to process the given roles" do
-      input_processor = instance_double(Rabarber::Input::Roles, process: roles)
-      allow(Rabarber::Input::Roles).to receive(:new).with(roles).and_return(input_processor)
-      expect(input_processor).to receive(:process)
-      subject
     end
   end
 
@@ -87,6 +76,14 @@ RSpec.describe Rabarber::HasRoles do
 
       it_behaves_like "it caches user roles", { context_type: "Project", context_id: nil }, [:admin, :developer]
     end
+
+    context "when given an invalid context" do
+      let(:context) { 123 }
+
+      it "raises with correct message" do
+        expect { subject }.to raise_error(Rabarber::InvalidArgumentError, "Expected an instance of ActiveRecord model, a Class, or nil, got 123")
+      end
+    end
   end
 
   describe "#all_roles" do
@@ -94,7 +91,7 @@ RSpec.describe Rabarber::HasRoles do
 
     let(:user) { User.create! }
 
-    shared_examples_for "it caches all user roles" do |all_roles|
+    shared_examples_for "it caches all user roles" do
       it "caches user roles" do
         expect(Rabarber::Core::Cache).to receive(:fetch).with([user.id, :all]) do |&block|
           result = block.call
@@ -108,7 +105,9 @@ RSpec.describe Rabarber::HasRoles do
     context "when the user has no roles" do
       it { is_expected.to eq({}) }
 
-      it_behaves_like "it caches all user roles", {}
+      it_behaves_like "it caches all user roles" do
+        let(:all_roles) { {} }
+      end
     end
 
     context "when the user has some roles" do
@@ -122,14 +121,18 @@ RSpec.describe Rabarber::HasRoles do
 
       it { is_expected.to eq(nil => [:admin, :manager], User => [:viewer], project => [:manager]) }
 
-      it_behaves_like("it caches all user roles", { nil => [:admin, :manager], User => [:viewer], Project.take => [:manager] })
+      it_behaves_like "it caches all user roles" do
+        let(:all_roles) { { nil => [:admin, :manager], User => [:viewer], project => [:manager] } }
+      end
 
       context "when the instance context can't be found" do
         before { project.destroy! }
 
         it { is_expected.to eq(nil => [:admin, :manager], User => [:viewer]) }
 
-        it_behaves_like "it caches all user roles", { nil => [:admin, :manager], User => [:viewer] }
+        it_behaves_like "it caches all user roles" do
+          let(:all_roles) { { nil => [:admin, :manager], User => [:viewer] } }
+        end
       end
 
       context "when the class context doesn't exist" do
@@ -159,7 +162,6 @@ RSpec.describe Rabarber::HasRoles do
     before { user.assign_roles(:admin, :manager) }
 
     it_behaves_like "role names are validated"
-    it_behaves_like "role names are processed"
 
     context "when the user has at least one of the given roles" do
       let(:roles) { [:admin, :accountant] }
@@ -190,6 +192,15 @@ RSpec.describe Rabarber::HasRoles do
 
       it { is_expected.to be true }
     end
+
+    context "when given an invalid context" do
+      let(:context) { 123 }
+      let(:roles) { [:admin] }
+
+      it "raises with correct message" do
+        expect { subject }.to raise_error(Rabarber::InvalidArgumentError, "Expected an instance of ActiveRecord model, a Class, or nil, got 123")
+      end
+    end
   end
 
   shared_examples_for "it deletes the cache" do |processed_context|
@@ -210,7 +221,6 @@ RSpec.describe Rabarber::HasRoles do
       let(:context) { Project }
 
       it_behaves_like "role names are validated"
-      it_behaves_like "role names are processed"
 
       context "when the given roles exist" do
         before do
@@ -298,7 +308,6 @@ RSpec.describe Rabarber::HasRoles do
       let(:context) { nil }
 
       it_behaves_like "role names are validated"
-      it_behaves_like "role names are processed"
 
       context "when the given roles exist" do
         let(:context) { Project }
@@ -384,6 +393,17 @@ RSpec.describe Rabarber::HasRoles do
         it { is_expected.to match_array(roles) }
       end
     end
+
+    context "when given an invalid context" do
+      let(:user) { User.create! }
+      let(:roles) { [:admin] }
+      let(:context) { 123 }
+      let(:create_new) { true }
+
+      it "raises with correct message" do
+        expect { subject }.to raise_error(Rabarber::InvalidArgumentError, "Expected an instance of ActiveRecord model, a Class, or nil, got 123")
+      end
+    end
   end
 
   describe "#revoke_roles" do
@@ -394,7 +414,6 @@ RSpec.describe Rabarber::HasRoles do
     let(:context) { nil }
 
     it_behaves_like "role names are validated"
-    it_behaves_like "role names are processed"
 
     context "when the user has the given roles" do
       let(:context) { Project }
@@ -443,6 +462,14 @@ RSpec.describe Rabarber::HasRoles do
 
       it { is_expected.to be_empty }
     end
+
+    context "when given an invalid context" do
+      let(:context) { 123 }
+
+      it "raises with correct message" do
+        expect { subject }.to raise_error(Rabarber::InvalidArgumentError, "Expected an instance of ActiveRecord model, a Class, or nil, got 123")
+      end
+    end
   end
 
   describe "#revoke_all_roles" do
@@ -488,6 +515,17 @@ RSpec.describe Rabarber::HasRoles do
           [user.id, :all]
         )
         subject
+      end
+    end
+
+    context "when the user has roles with an invalid context key" do
+      before do
+        user.assign_roles(:admin)
+        allow(user).to receive(:all_roles).and_return({ 123 => [:admin] })
+      end
+
+      it "raises with correct message" do
+        expect { subject }.to raise_error(Rabarber::InvalidArgumentError, "Expected an instance of ActiveRecord model, a Class, or nil, got 123")
       end
     end
   end