+++ spec/support/configuration.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Rabarber
+  module Configuration
+    enable_test_interface
+  end
+end