+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v5.1.1
+
+### Misc:
+
+- Switched to `dry-types` for input validation
+- Switched to `dry-configurable` for configuration management
+
 ## v5.1.0
 
 ### Features: